+++ apps/studio/components/layouts/SQLEditorLayout/SQLEditorMenu.tsx
@@ -8,7 +8,7 @@ import { toast } from 'sonner'
 
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
 import { useLocalStorage } from 'hooks/misc/useLocalStorage'
-import { useSelectedProject } from 'hooks/misc/useSelectedProject'
+import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
 import { useProfile } from 'lib/profile'
 import { getAppStateSnapshot } from 'state/app-state'
 import { useSqlEditorV2StateSnapshot } from 'state/sql-editor-v2'
@@ -28,15 +28,14 @@ import {
   InnerSideBarFilterSortDropdown,
   InnerSideBarFilterSortDropdownItem,
 } from 'ui-patterns/InnerSideMenu'
-import { SqlEditorMenuStaticLinks } from './SqlEditorMenuStaticLinks'
 import { SearchList } from './SQLEditorNavV2/SearchList'
 import { SQLEditorNav } from './SQLEditorNavV2/SQLEditorNav'
 
 export const SQLEditorMenu = () => {
   const router = useRouter()
-  const { profile } = useProfile()
-  const project = useSelectedProject()
   const { ref } = useParams()
+  const { profile } = useProfile()
+  const { data: project } = useSelectedProjectQuery()
   const snapV2 = useSqlEditorV2StateSnapshot()
 
   const [search, setSearch] = useState('')
@@ -153,14 +152,7 @@ export const SQLEditorMenu = () => {
           </DropdownMenu>
         </div>
 
-        {showSearch ? (
-          <SearchList search={debouncedSearch} />
-        ) : (
-          <>
-            <SQLEditorNav sort={sort} />
-            <SqlEditorMenuStaticLinks />
-          </>
-        )}
+        {showSearch ? <SearchList search={debouncedSearch} /> : <SQLEditorNav sort={sort} />}
       </div>
 
       <div className="p-4 border-t sticky bottom-0 bg-studio">
