+++ apps/studio/components/layouts/SQLEditorLayout/SQLEditorNavV2/SQLEditorNav.constants.ts
@@ -1,0 +1,13 @@
+export type SectionState = {
+  shared: boolean
+  favorite: boolean
+  private: boolean
+  community: boolean
+}
+
+export const DEFAULT_SECTION_STATE: SectionState = {
+  shared: false,
+  favorite: false,
+  private: true,
+  community: true,
+}
